Restructure page files and add Movie type
