+++ src/index.css
@@ -1,0 +1,3 @@
+.section {
+  margin: 0 1rem;
+}
