--- src/pages/FilmView.tsx
@@ -1,7 +1,0 @@
-export default function FilmView() {
-  return (
-    <div>
-      <h1>FilmView</h1>
-    </div>
-  );
-}
