+++ src/pages/FilmView/FilmView.tsx
@@ -1,0 +1,10 @@
+// import { useParams } from "react-router-dom";
+// import films from "../data/movies.json"
+
+export default function FilmView() {
+  // const { slug } = useParams<{ slug: string }>();
+
+  // const { title, year, rating, actors, genre, synopsis, thumbnail } = JSON.parse(films).find((film: any) => film.slug === slug);
+
+  return <div className='section' />;
+}
