+++ src/types/film.d.ts
@@ -1,0 +1,11 @@
+export default interface Movie {
+  title: string;
+  slug: string;
+  id: number;
+  year: number;
+  rating: string;
+  actors: string[];
+  genre: string;
+  synopsis: string;
+  thumbnail: string;
+}
